requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pydantic>=2.0.0
